+++ ecs/src/v2/EcsClient.cpp
@@ -1795,46 +1795,6 @@ std::shared_ptr<ShowServerTagsResponse> EcsClient::showServerTags(ShowServerTags
 
     return localVarResult;
 }
-std::shared_ptr<UpdateAutoTerminateTimeServerResponse> EcsClient::updateAutoTerminateTimeServer(UpdateAutoTerminateTimeServerRequest &request)
-{
-    std::string localVarPath = "/v1/{project_id}/cloudservers/{server_id}/actions/update-auto-terminate-time";
-
-    std::map<std::string, std::string> localVarQueryParams;
-    std::map<std::string, std::string> localVarHeaderParams;
-    std::map<std::string, std::string> localVarFormParams;
-    std::map<std::string, std::string> localVarPathParams;
-    std::map<std::string, std::shared_ptr<HttpContent>> localVarFileParams;
-
-    localVarPathParams["server_id"] = parameterToString(request.getServerId());
-
-    bool isJson = false;
-    bool isMultiPart = false;
-    std::string contentType = getContentType("application/json;charset=UTF-8", isJson, isMultiPart);
-    localVarHeaderParams["Content-Type"] = contentType;
-
-
-    std::string localVarHttpBody;
-    if (isJson) {
-        web::json::value localVarJson;
-        localVarJson = ModelBase::toJson(request.getBody());
-        localVarHttpBody = utility::conversions::to_utf8string(localVarJson.serialize());
-    }
-
-    std::unique_ptr<HttpResponse> res = callApi("POST", localVarPath, localVarPathParams, localVarQueryParams, localVarHeaderParams, localVarHttpBody);
-
-    std::shared_ptr<UpdateAutoTerminateTimeServerResponse> localVarResult = std::make_shared<UpdateAutoTerminateTimeServerResponse>();
-
-    if (!res->getHttpBody().empty()) {
-        utility::string_t localVarResponse = utility::conversions::to_string_t(res->getHttpBody());
-        web::json::value localVarJson = web::json::value::parse(localVarResponse);
-        localVarResult->fromJson(localVarJson);
-    }
-    localVarResult->setStatusCode(res->getStatusCode());
-    localVarResult->setHeaderParams(res->getHeaderParams());
-    localVarResult->setHttpBody(res->getHttpBody());
-
-    return localVarResult;
-}
 std::shared_ptr<UpdateServerResponse> EcsClient::updateServer(UpdateServerRequest &request)
 {
     std::string localVarPath = "/v1/{project_id}/cloudservers/{server_id}";
@@ -1863,6 +1823,46 @@ std::shared_ptr<UpdateServerResponse> EcsClient::updateServer(UpdateServerReques
     std::unique_ptr<HttpResponse> res = callApi("PUT", localVarPath, localVarPathParams, localVarQueryParams, localVarHeaderParams, localVarHttpBody);
 
     std::shared_ptr<UpdateServerResponse> localVarResult = std::make_shared<UpdateServerResponse>();
+
+    if (!res->getHttpBody().empty()) {
+        utility::string_t localVarResponse = utility::conversions::to_string_t(res->getHttpBody());
+        web::json::value localVarJson = web::json::value::parse(localVarResponse);
+        localVarResult->fromJson(localVarJson);
+    }
+    localVarResult->setStatusCode(res->getStatusCode());
+    localVarResult->setHeaderParams(res->getHeaderParams());
+    localVarResult->setHttpBody(res->getHttpBody());
+
+    return localVarResult;
+}
+std::shared_ptr<UpdateServerAutoTerminateTimeResponse> EcsClient::updateServerAutoTerminateTime(UpdateServerAutoTerminateTimeRequest &request)
+{
+    std::string localVarPath = "/v1/{project_id}/cloudservers/{server_id}/actions/update-auto-terminate-time";
+
+    std::map<std::string, std::string> localVarQueryParams;
+    std::map<std::string, std::string> localVarHeaderParams;
+    std::map<std::string, std::string> localVarFormParams;
+    std::map<std::string, std::string> localVarPathParams;
+    std::map<std::string, std::shared_ptr<HttpContent>> localVarFileParams;
+
+    localVarPathParams["server_id"] = parameterToString(request.getServerId());
+
+    bool isJson = false;
+    bool isMultiPart = false;
+    std::string contentType = getContentType("application/json;charset=UTF-8", isJson, isMultiPart);
+    localVarHeaderParams["Content-Type"] = contentType;
+
+
+    std::string localVarHttpBody;
+    if (isJson) {
+        web::json::value localVarJson;
+        localVarJson = ModelBase::toJson(request.getBody());
+        localVarHttpBody = utility::conversions::to_utf8string(localVarJson.serialize());
+    }
+
+    std::unique_ptr<HttpResponse> res = callApi("POST", localVarPath, localVarPathParams, localVarQueryParams, localVarHeaderParams, localVarHttpBody);
+
+    std::shared_ptr<UpdateServerAutoTerminateTimeResponse> localVarResult = std::make_shared<UpdateServerAutoTerminateTimeResponse>();
 
     if (!res->getHttpBody().empty()) {
         utility::string_t localVarResponse = utility::conversions::to_string_t(res->getHttpBody());