+++ ecs/src/v2/model/UpdateServerAutoTerminateTimeResponse.cpp
@@ -1,0 +1,46 @@
+
+
+#include "huaweicloud/ecs/v2/model/UpdateServerAutoTerminateTimeResponse.h"
+
+namespace HuaweiCloud {
+namespace Sdk {
+namespace Ecs {
+namespace V2 {
+namespace Model {
+
+
+
+
+UpdateServerAutoTerminateTimeResponse::UpdateServerAutoTerminateTimeResponse()
+{
+}
+
+UpdateServerAutoTerminateTimeResponse::~UpdateServerAutoTerminateTimeResponse() = default;
+
+void UpdateServerAutoTerminateTimeResponse::validate()
+{
+}
+
+web::json::value UpdateServerAutoTerminateTimeResponse::toJson() const
+{
+    web::json::value val = web::json::value::object();
+
+
+    return val;
+}
+
+bool UpdateServerAutoTerminateTimeResponse::fromJson(const web::json::value& val)
+{
+    bool ok = true;
+    
+    return ok;
+}
+
+
+}
+}
+}
+}
+}
+
+