+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 3.0.4-beta 2021-01-30
+## HuaweiCloud SDK ECS
+- ### Features
+    - None
+- ### Bug Fix
+    - None
+- ### Change
+    - Change interface name from `UpdateAutoTerminateTimeServer` to `UpdateServerAutoTerminateTime`.
+
+## HuaweiCloud SDK EVS
+- ### Features
+    - None
+- ### Bug Fix
+    - None
+- ### Change
+    - Interface `CinderCreateVolume` is supported to specify the id of dedicated storage pool using property `OS-SCH-HNT:scheduler_hints`.
+    - Modify property type of `allocated` of class `QuotaDetails` from `String` to `Integer`.
+
+
 # 3.0.3-beta 2021-01-25
 ## HuaweiCloud SDK Core
 - ### Features